+++ CHANGELOG.md
@@ -33,6 +33,7 @@ For notes on migrating to 2.x / 0.200.x see [the upgrade guide](doc/upgrade-to-2
 ### :house: Internal
 
 * perf(sdk-trace-base): optimize TraceIdRatioBasedSampler performance [#6284](https://github.com/open-telemetry/opentelemetry-js/pull/6284) @AbhiPrasad
+* test: test Node.js 26 in CI [#6671](https://github.com/open-telemetry/opentelemetry-js/pull/6671) @cjihrig
 
 ## 2.7.0